requirements.txt
# Berdasarkan settings.py
django==5.2.4
psycopg2-binary
djangorestframework
django-filter
django-cors-headers